move io-funtions into io.rs
